Internal functions for pulling apart fitContinuous models
Generic functions for extracting the model name (e.g., BM, OU), the
underlying data, the fitted parameters, and an object that combines
all of these that will be useful when making a unit tree.

None of these are exported, and should be seen as liable to change, as
we work out what we really need.  I don't see that the user should ever
need to care what is going on here though.

This relies on the current version of geiger, so updated DESCRIPTION to
reflect this.