Update unit tests to work with recent diversitree.
Fixes #14

I really don't know why this is failing on recent versions and not on
older versions.  I don't see any changes that might have caused it
aside from possibly richfitz/diversitree@f2845b6 but that doesn't
seem likely.  In any case, the fix is easy and it's in a bit of code
that was a workaround for testing anyway.